requirements.txt
Django==5.0.6
django-widget-tweaks==1.5.0
PyYAML==6.0.1
dt4dds==1.1.0